Fix "Response from service dropped" boxes.
Cap the number of messages that can be in-flight to the Pebble at once.
Once ten messages are in flight, enqueue any further messages until
ACKs or NACKs are received.

Sending messages is still unreliable - if the pebble drops a message, we
will silently send the next one (and the user gets a bit of annoying UI).

Holding messages for transmission until the previous one is complete
results in excessively slow message delivery.

Signed-off-by: Katharine Berry <ktbry@google.com>